+++ src/scicloj/clay/v2/make.clj
@@ -371,7 +371,7 @@
       (catch Throwable e
         (-> spec
             (assoc :page (-> spec
-                             (assoc :items [(item/pprint e)])
+                             (assoc :items [(item/print-throwable e)])
                              page/html))
             server/update-page!)
         (throw e))